+++ packages/infinity-modules/production-log/src/views/ProductionMachineView.vue
@@ -1,50 +1,38 @@
 <template>
   <div>
-    <portal to="app-extension">
-      <v-tabs
-        dense
-        center-active
-        v-model="logView"
-      >
-        <v-tab class="text-none">
-          Production
-        </v-tab>
-        <v-tab class="text-none">
-          Downtime
-        </v-tab>
-        <!-- <v-tab class="text-none">
-          Operators
-        </v-tab> -->
-      </v-tabs>
-    </portal>
-    <production-log-production-view v-if="logView === 0" />
-    <production-log-downtime-view v-else-if="logView === 1" />
-    <production-log-operator-view v-if="logView === 2" />
+    <template v-if="loading">
+      Loading...
+    </template>
+    <dashboard-toolbar-extension/>
+    <production-log-production-view/>
   </div>
 </template>
 
 <script>
-import { mapMutations } from 'vuex';
+import { mapMutations, mapActions } from 'vuex';
+import DashboardToolbarExtension from '../components/core/DashboardToolbarExtension.vue';
 import ProductionLogProductionView from '../components/ProductionLogProductionView.vue';
-import ProductionLogDowntimeView from '../components/ProductionLogDowntimeView.vue';
-import ProductionLogOperatorView from '../components/ProductionLogOperatorView.vue';
+
 
 export default {
   name: 'ProductionMachineView',
   components: {
-    ProductionLogOperatorView,
     ProductionLogProductionView,
-    ProductionLogDowntimeView,
+    DashboardToolbarExtension,
   },
   data() {
     return {
-      logView: 0,
+      loading: false,
     };
   },
-  created() {
-    const view = localStorage.getItem('logView');
-    this.logView = view ? JSON.parse(view) : 0;
-    this.setExtendedHeader(true);
+  async created() {
+    await this.fetchBusinessHours();
+    await this.fetchRejectionReasons();
+    const machines = await this.fetchMachines();
+    if (machines) {
+      this.setExtendedHeader(true);
+    }
+    this.loading = false;
   },
   watch: {
     logView(val) {
@@ -53,6 +41,7 @@ export default {
   },
   methods: {
     ...mapMutations('helper', ['setExtendedHeader']),
+    ...mapActions('productionLog', ['fetchMachines', 'fetchBusinessHours', 'fetchRejectionReasons']),
   },
 };
 </script>